moved function for generate identifierCode to new
