+++ cypress/support/utils/dateTools.js
@@ -83,17 +83,6 @@ export default {
     return (getRandomStringCode(4)) + Math.floor((Math.random() * (Math.floor(max) - Math.ceil(min)) + Math.ceil(min))).toString();
   },
 
-  getRandomIdentifierCode: () => {
-    // returns random 3 digit code for the identifier in the marcBib file
-    const lettersUC = 'ABDEFGHKM';
-    let str = '';
-    for (let i = 0; i < 3; i++) {
-      const position = Math.floor(Math.random() * lettersUC.length);
-      str += lettersUC.substring(position, position + 1);
-    }
-    return str;
-  },
-
   getDateRanges: (rangesCount = 1, isInDifferentYears = false) => {
     const resultRanges = [];
     const currentDate = new Date();
